requirements.txt
torch==2.0.0
torchvision==0.15.1
numpy==1.23.5
opencv-python==4.8.0.74
opencv-python-headless==4.8.0.74
Pillow==9.5.0
tqdm==4.66.1
tensorboard==2.13.0